Finished up the user controller.

+++ drone_api/src/AppBundle/Controller/UsersController.php
@@ -27,24 +27,24 @@ class UsersController extends BasicApiController{
      * @Route("/user/{id}", name="user")
      * @Method({"GET"})
      * @param $id
+     * @return Response
      */
-    public function showAction($id = 0){
-        $response = $this->getStandardResponse();
-        if ($id != 0) {
-            $repository = $this->getDoctrine()->getRepository('AppBundle:Users');
+    public function showAction($id = -1){
+        $response = $this->getStandardResponseFormat();
+        if ($id > -1) {
+            $repository = $this->getDoctrine()->getRepository(Strings::$APPBUNDLE_USER);
             $user = $repository->find($id);
             if ($user) {
-                $responseParams = array(Strings::$MESSAGE=>Strings::$OK, Strings::$STATUS=>Strings::$STATUS_FOUND);
-                $response->setContent(json_encode($this->mergeData($responseParams,json_encode($user))));
-                return $response;
+                $responseParams = array(Strings::$MESSAGE=>Strings::$OK, Strings::$STATUS=>Strings::$STATUS_OK);
+                $response->setContent($this->getStandardJSONResponse($responseParams,$user,Strings::$USER));
             }else{
                 $responseParams = array(Strings::$MESSAGE=>Strings::$MESSAGE_COULD_NOT_FIND_USER, Strings::$STATUS=>Strings::$STATUS_NOT_FOUND);
-                $response->setContent(json_encode($responseParams,$response->getContent()));
-                return $response;
+                $response->setContent(json_encode($responseParams));
             }
+            return $response;
         }
         $responseParams = array(Strings::$MESSAGE=>Strings::$MISSING_PARAMS, Strings::$STATUS=>Strings::$STATUS_BAD_REQUEST);
-        $response->setContent(json_encode($this->mergeData($responseParams,$response->getContent())));
+        $response->setContent($responseParams);
         return $response;
     }
 
@@ -56,7 +56,7 @@ class UsersController extends BasicApiController{
      * @internal param $data
      */
     public function createAction(Request $request){
-        $response = $this->getStandardResponse();
+        $response = $this->getStandardResponseFormat();
         $userJSON = $this->getParamsInContent($request,Strings::$USER);
             if ($this->requiredRequestContent(array(Strings::$USER_PASSWORD,Strings::$USER_EMAIL,Strings::$USER_FIRST_NAME),$userJSON)) {
                 $user = new Users();
@@ -65,51 +65,69 @@ class UsersController extends BasicApiController{
                 $em->persist($user);
                 $em->flush();
                 $responseParams = array(Strings::$MESSAGE=>Strings::$OK, Strings::$STATUS=>Strings::$STATUS_OK);
-                $response->setContent(json_encode($this->mergeData($userJSON, $responseParams)));
+                $response->setContent($this->getStandardJSONResponse($responseParams,$user,Strings::$USER));
                 return $response;
         }
         $responseParams = array(Strings::$MESSAGE=>Strings::$MISSING_PARAMS, Strings::$STATUS=>Strings::$STATUS_BAD_REQUEST);
-        $response->setContent(json_encode($this->mergeData($responseParams,$response->getContent())));
+        $response->setContent($this->getStandardJSONResponse($responseParams,$response->getContent()));
         return $response;
     }
 
     /**
-     * @Route("/user/edit")
+     * @Route("/user/update")
      * @Method({"PUT"})
-     * @param $data
+     * @param Request $request
+     * @return Response|void
+     * @internal param $data
      */
     public function updateAction(Request $request){
-        $response = $this->getStandardResponse();
+        $response = $this->getStandardResponseFormat();
         $user = $this->getParamsInContent($request,Strings::$USER);
-        if (property_exists($user,Strings::$USER_ID)) {
+        if (array_key_exists(Strings::$USER_ID,$user)) {
             $em = $this->getDoctrine()->getManager();
-            $foundUser = $em->getRepository('AppBundle:Users')->find($user[Strings::$USER_ID]);
+            $foundUser = $em->getRepository(Strings::$APPBUNDLE_USER)->find($user[Strings::$USER_ID]);
             if ($foundUser){
                 $foundUser->setObject($user);
+                $em->flush();
+                $responseParams = array(Strings::$MESSAGE=>Strings::$OK, Strings::$STATUS=>Strings::$STATUS_OK);
+                $response->setContent($this->getStandardJSONResponse($responseParams,$user));
+            }else{
+                $responseParams = array(Strings::$MESSAGE=>Strings::$MESSAGE_COULD_NOT_FIND_USER, Strings::$STATUS=>Strings::$STATUS_NOT_FOUND);
+                $response->setContent(json_encode($responseParams));
             }
-            $em->flush();
+            return $response;
         }
+        $responseParams = array(Strings::$MESSAGE=>Strings::$MISSING_PARAMS, Strings::$STATUS=>Strings::$STATUS_BAD_REQUEST);
+        $response->setContent($this->getStandardJSONResponse($responseParams,$response->getContent()));
+        return $response;
     }
 
     /**
      * @Route("/user/delete")
      * @Method({"DELETE"})
-     * @param $id
+     * @param Request $request
+     * @return Response
+     * @internal param $id
      */
     public function deleteAction(Request $request){
-        $response = $this->getStandardResponse();
+        $response = $this->getStandardResponseFormat();
         $em = $this->getDoctrine()->getManager();
         $user = $this->getParamsInContent($request,Strings::$USER);
-        if (property_exists($user,Strings::$USER_ID)){
-            $foundUser = $em->getRepository('AppBundle:Users')->find($user[Strings::$USER_ID]);
-            $em->remove($foundUser);
-            $em->flush();
-            $responseParams = array(Strings::$MESSAGE=>Strings::$MESSAGE_DELETED_USER, Strings::$STATUS=>Strings::$STATUS_OK);
-            $response->setContent(json_encode($responseParams));
+        if (property_exists($user,Strings::$USER_ID)) {
+            $foundUser = $em->getRepository(Strings::$APPBUNDLE_USER)->find($user[Strings::$USER_ID]);
+            if ($foundUser) {
+                $em->remove($foundUser);
+                $em->flush();
+                $responseParams = array(Strings::$MESSAGE => Strings::$MESSAGE_DELETED_USER, Strings::$STATUS => Strings::$STATUS_OK);
+                $response->setContent($this->getStandardJSONResponse($responseParams, $foundUser));
+            } else {
+
+            }
             return $response;
         }
+
         $responseParams = array(Strings::$MESSAGE=>Strings::$MISSING_PARAMS, Strings::$STATUS=>Strings::$STATUS_BAD_REQUEST);
-        $response->setContent(json_encode($this->mergeData($responseParams,$response->getContent())));
+        $response->setContent($responseParams);
         return $response;
     }
